Revert "update test"
This reverts commit d81333849d7a2647017473beed4b7690cbccfd36.

+++ tests/acceptance/test_files/service_and_plugin.bats
@@ -1,6 +1,5 @@
 load "$LIB_BATS_ASSERT/load.bash"
 load "$LIB_BATS_SUPPORT/load.bash"
-load "$LIB_BATS_FILE/load.bash"
 
 @test "steampipe plugin help is displayed when no sub command given" {
   steampipe plugin > test.txt
@@ -545,8 +544,8 @@ load "$LIB_BATS_FILE/load.bash"
   vFile1="$tmpdir/plugins/hub.steampipe.io/plugins/turbot/net@latest/version.json"
   vFile2="$tmpdir/plugins/hub.steampipe.io/plugins/turbot/chaos@latest/version.json"
   
-  assert_file_exists $vFile1
-  assert_file_exists $vFile2
+  [ ! -f $vFile1 ] && fail "could not find $vFile1"
+  [ ! -f $vFile2 ] && fail "could not find $vFile2"
   
   rm -rf $tmpdir
 }
@@ -569,8 +568,8 @@ load "$LIB_BATS_FILE/load.bash"
   # run steampipe again so that the plugin version files get backfilled
   run steampipe query "select 1" --install-dir $tmpdir
   
-  assert_exists $vFile1
-  assert_exists $vFile2
+  [ ! -f $vFile1 ] && fail "could not find $vFile1"
+  [ ! -f $vFile2 ] && fail "could not find $vFile2"
   
   assert_equal "$(cat $vFile1)" "$file1Content"
   assert_equal "$(cat $vFile2)" "$file2Content"
@@ -595,8 +594,8 @@ load "$LIB_BATS_FILE/load.bash"
   # run steampipe again so that the plugin version files get backfilled
   run steampipe query "select 1" --install-dir $tmpdir
   
-  assert_exists $vFile1
-  assert_exists $vFile2
+  [ ! -f $vFile1 ] && fail "could not find $vFile1"
+  [ ! -f $vFile2 ] && fail "could not find $vFile2"
   
   assert_equal "$(cat $vFile1)" "$file1Content"
   assert_equal "$(cat $vFile2)" "$file2Content"
@@ -604,7 +603,7 @@ load "$LIB_BATS_FILE/load.bash"
   rm -rf $tmpdir
 }
 
-@test "verify that global plugin/versions.json is composed from individual version.json files when global is absent" {
+@test "verify that global plugin/versions.json is composed from individual version.json files when it is absent" {
   tmpdir=$(mktemp -d)
   run steampipe plugin install net chaos --install-dir $tmpdir
   assert_success
@@ -619,7 +618,7 @@ load "$LIB_BATS_FILE/load.bash"
   # run steampipe again so that the plugin version files get backfilled
   run steampipe query "select 1" --install-dir $tmpdir
   
-  assert_file_exists $vFile
+  [ ! -f $vFile ] && fail "could not find $vFile"
   
   assert_equal "$(cat $vFile)" "$fileContent"
   
@@ -640,7 +639,7 @@ load "$LIB_BATS_FILE/load.bash"
   # run steampipe again so that the plugin version files get backfilled
   run steampipe query "select 1" --install-dir $tmpdir
   
-  assert_file_exists $vFile
+  [ ! -f $vFile ] && fail "could not find $vFile"
   
   assert_equal "$(cat $vFile)" "$fileContent"
   
@@ -665,7 +664,7 @@ load "$LIB_BATS_FILE/load.bash"
   run steampipe query "select 1" --install-dir $tmpdir
 
   # verify that global file got created
-  assert_file_exists $vFile
+  [ ! -f $vFile ] && fail "could not find $vFile"
   
   rm -rf $tmpdir
 }
